+++ src/clj/skylobby/server.clj
@@ -200,7 +200,9 @@
                                           :engine-overrides engine-overrides
                                           :host-ingame true
                                           :spring-isolation-dir spring-root}
-                                         (resource/spring-root-resources spring-root by-spring-root))))
+                                         (dissoc
+                                           (resource/spring-root-resources spring-root by-spring-root)
+                                           :engine-version :map-name :mod-name))))
       (defmethod -event-msg-handler
         :skylobby/send-message
         [{:keys [?data]}]